requirements.txt
BOTO3
Langchain
Pydantic
python-dotenv
pandas